fix: Phase 2 finalization — complete InvoiceStatus enum migration across all consumers
Yesterday's Phase 2 commit added `'status' => InvoiceStatus::class`
to Invoice's $casts, but most consumers still compared `$invoice->status`
to STRING literals like 'paid', 'voided', 'cancelled'. After the cast
took effect, $invoice->status is an enum instance, so:

  in_array($enum, ['paid', ...]) → false
  whereIn('status', ['paid', ...]) → empty result

This commit migrates every comparison/whereIn/scope across:
- Models (PaymentLink, Payment, ArchivedPayment, NotificationTemplate)
- Policies (Invoice, Building, Invitation)
- Services (PaymentLinkService, InvoiceService, ImportService,
  FinanceExport/ReportService, ReconciliationService, etc.)
- Controllers (Invoice, CreditNote, Notification, Onboarding, plus API)
- Observers (Expense, Invoice, Refund)
- Form Requests (BulkOperations, Building, MoveOut, Kyc, etc.)
- Migrations (factories paired with new untracked migrations)
- Tests paired with the above (asserting enum semantics)

Also picks up several non-enum WIP fixes that were sitting in the
working tree:
- Auth/Registration request validation tightening
- Mail template branding via config('app.name')
- DepositTransformer field naming
- Domain/Paystack exception handling improvements
- BankReconciliation, AfricasTalking, Scheduler service hardening

+++ tests/Traits/MocksExternalServices.php
@@ -42,7 +42,7 @@ trait MocksExternalServices
         $this->app->instance(PaystackService::class, $mock);
     }
 
-    protected function getMockMpesaStkSuccessCallback(string $checkoutRequestId, int $amount, string $phone): array
+    protected function getMockMpesaStkSuccessCallback(string $checkoutRequestId, float $amount, string $phone): array
     {
         return [
             'Body' => [
